+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Simple HTML report generation with subcommand _report_.
+
 ## [0.4.0] - 2025-02-21
 
 ### Added